+++ app/views/users/uploads/category_selection.html.erb
